renamed routers and put them into routes folder. renmaed variables to handle potential query errors and to match names in both databases
