+++ public/setbudget.html
@@ -46,7 +46,7 @@
             <a href="expenses.html"><img src="./img/greenarrow.jpg"></a>
         </div>
         <div class="inputs">
-            <form action="/setValues/setMonthlySpend" method="POST">
+            <form action="/budget" method="POST">
                 <label for="mortgageRent">Mortgage/Rent $</label>
                 <input type="number" name="mortgageRent" id="mortageRent">
                 <br>
